breaking: move execution.msc into legacy
