+++ mxops/execution/legacy_steps.py
@@ -42,7 +42,7 @@ from mxops.execution import utils
 
 from mxops.execution.account import AccountsManager
 from mxops.execution.legacy_checks import Check, SuccessCheck, instanciate_checks
-from mxops.execution.msc import EsdtTransfer
+from mxops.execution.legacy_msc import EsdtTransfer
 from mxops.execution.network import send, send_and_wait_for_result
 from mxops.utils.logger import get_logger
 from mxops.utils.msc import get_file_hash, get_tx_link
